requirements.txt
pyyaml
pytest
pytest-cov
black
flake8
mypy
watchdog
rich
types-PyYAML
